app/pyproject.toml
[tool.poetry]
name = "chatty"
version = "0.1.0"
description = "Chatty Backend experimentation"
authors = ["Arturo Boston <person@person.com>"]
packages = [{include = "chatty", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
requests = "^2.32.5"
structlog = "^23.2.0"
python-socketio = "^5.10.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"
python-socketio = {extras = ["client"], version = "^5.10.0"}
aiohttp = "^3.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/alembic/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
skip_covered = false
precision = 2
fail_under = 70.0

[tool.coverage.html]
directory = "htmlcov"

